Clean up code, delete Main activity because redundant
